Modified design of the application, fixed minor bugs, lost the will to live and regained it afterwards ^^

+++ app/src/main/java/com/tkkd/mealplanner/Database/Entities/Home.java
@@ -13,7 +13,7 @@ public class Home {
     @PrimaryKey(autoGenerate = true)
     public long id;
 
-    public int quantity;
+    public float quantity;
 
     @ColumnInfo(name = "ingredient_id")
     public long ingId;
